optimize async extension read/write with bytemuck

+++ programs/async_vault/src/extensions/pausable_redemptions/instructions/initialize_pausable_redemptions.rs
@@ -38,7 +38,7 @@ pub fn handler(
         &ctx.accounts.vault.to_account_info(),
         &ctx.accounts.vault,
         &PausableRedemption {
-            paused: args.paused,
+            paused: args.paused as u8,
         },
     )
 }
